+++ examples/quote_spider.py
@@ -3,8 +3,7 @@
 from urllib.parse import urljoin
 import json
 
-from xpaw import Spider, HttpRequest, Selector
-from xpaw.run import run_spider
+from xpaw import Spider, HttpRequest, Selector, run_spider
 
 
 class QuotesSpider(Spider):