+++ examples/xpath_quote_spider.py
@@ -2,8 +2,7 @@
 
 from urllib.parse import urljoin
 
-from xpaw import Spider, HttpRequest, Selector
-from xpaw.run import run_spider
+from xpaw import Spider, HttpRequest, Selector, run_spider
 
 
 class XPathQuotesSpider(Spider):