+++ xpaw/__init__.py
@@ -5,13 +5,19 @@ from . import _patch
 
 from aiohttp import FormData
 
-from .spider import Spider, every
 from .http import HttpRequest, HttpResponse
-from .selector import Selector
 from .downloader import Downloader
+from .spider import Spider, every
+from .selector import Selector
 from .item import Item, Field
+from .run import run_spider, run_crawler
 
-__all__ = ('FormData', 'HttpRequest', 'HttpResponse', 'Downloader', 'Spider', 'Selector',
-           'Item', 'Field', 'every')
+__all__ = ('FormData',
+           'HttpRequest', 'HttpResponse',
+           'Downloader',
+           'Spider', 'every',
+           'Selector',
+           'Item', 'Field',
+           'run_spider', 'run_crawler')
 
 del _patch